pyproject.toml
[build-system]
requires = [
    "setuptools<69",
    "wheel<0.41"
]
build-backend = "setuptools.build_meta"


[project]
name = "microsys"
version = "1.0.0"
description = "micro (System Integration Services) - RTL Django Starter Pack, with Users Management, Permissions, Auto Section Management Page, and Activity Logging, and more."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "DeBeski", email = "debeski1@gmail.com" }
]
keywords = ["django", "microsys", "permissions", "authentication"]

# Dependencies
dependencies = [
    "Django>=5.1",
    "django-crispy-forms>=2.4",
    "django-tables2>=2.7",
    "django-filter>=24.3",
    "pillow>=11.0",
    "babel>=2.1",
]

# Classifiers
classifiers = [
    "Framework :: Django",
    "Framework :: Django :: 5",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# Tell setuptools which packages to include
[tool.setuptools.packages.find]
include = ["microsys"]

# Include static/templates files (equivalent to include_package_data=True)
[tool.setuptools]
include-package-data = true

[project.urls]
Homepage = "https://github.com/debeski/microsys"